move rmn_home to the pkg
